refactor script cmd
