+++ bin/opt8n/src/opt8n.rs
@@ -7,6 +7,7 @@ use anvil::{cmd::NodeArgs, eth::EthApi, NodeConfig, NodeHandle};
 use anvil_core::eth::block::Block;
 use anvil_core::eth::transaction::PendingTransaction;
 use cast::traces::{GethTraceBuilder, TracingInspectorConfig};
+use clap::Parser;
 use forge_script::ScriptArgs;
 use std::{
     error::Error,
@@ -14,7 +15,6 @@ use std::{
     path::PathBuf,
 };
 
-use clap::{CommandFactory, FromArgMatches, Parser};
 use color_eyre::eyre::{ensure, eyre, Result};
 use futures::StreamExt;
 use op_test_vectors::execution::{ExecutionFixture, ExecutionReceipt, ExecutionResult};
@@ -23,8 +23,6 @@ use revm::{
     primitives::{BlobExcessGasAndPrice, BlockEnv, CfgEnv, Env, SpecId, U256},
     Database, DatabaseCommit, DatabaseRef, Evm, EvmBuilder,
 };
-use serde::{Deserialize, Serialize};
-use tokio::io::{AsyncBufReadExt, BufReader};
 
 #[derive(Parser, Clone, Debug)]
 pub struct Opt8nArgs {
@@ -84,81 +82,6 @@ impl Opt8n {
             node_config,
             output_file,
         })
-    }
-
-    /// Run a Forge script with the given arguments, and generate an execution fixture
-    /// from the broadcasted transactions.
-    pub async fn run_script(self, script_args: Box<ScriptArgs>) -> Result<()> {
-        let mut new_blocks = self.eth_api.backend.new_block_notifications();
-
-        // Run the forge script and broadcast the transactions to the anvil node
-        let mut opt8n = self.broadcast_transactions(script_args).await?;
-
-        // Mine the block and generate the execution fixture
-        opt8n.mine_block().await;
-
-        let block = new_blocks.next().await.ok_or(eyre!("No new block"))?;
-        if let Some(block) = opt8n.eth_api.backend.get_block_by_hash(block.hash) {
-            opt8n.generate_execution_fixture(block).await?;
-        }
-
-        Ok(())
-    }
-
-    async fn broadcast_transactions(self, script_args: Box<ScriptArgs>) -> Result<Self> {
-        // Run the script, compile the transactions and broadcast to the anvil instance
-        let compiled = script_args.preprocess().await?.compile()?;
-
-        let pre_simulation = compiled
-            .link()
-            .await?
-            .prepare_execution()
-            .await?
-            .execute()
-            .await?
-            .prepare_simulation()
-            .await?;
-
-        let bundled = pre_simulation.fill_metadata().await?.bundle().await?;
-
-        let tx_count = bundled
-            .sequence
-            .sequences()
-            .iter()
-            .fold(0, |sum, sequence| sum + sequence.transactions.len());
-
-        // TODO: break into function
-        let broadcast = bundled.broadcast();
-
-        let opt8n = self;
-
-        let pending_transactions = tokio::task::spawn(async move {
-            loop {
-                let pending_tx_count = opt8n
-                    .eth_api
-                    .txpool_content()
-                    .await
-                    .expect("Failed to get txpool content")
-                    .pending
-                    .len();
-
-                if pending_tx_count == tx_count {
-                    return opt8n;
-                }
-            }
-        });
-
-        let opt8n = tokio::select! {
-            _ = broadcast => {
-                // TODO: Gracefully handle this error
-                return Err(eyre!("Script failed early"));
-            },
-            opt8n = pending_transactions => {
-                opt8n?
-            }
-        };
-
-        Ok(opt8n)
     }
 
     /// Updates the pre and post state allocations of the [ExecutionFixture] from Revm.
